CGO_ENABLED=1

+++ docker/build.sh
@@ -1,7 +1,11 @@
 #/bin/sh
+set -e
 IMAGE_TAG=filesync-web:latest
-rm .dist -r
-CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o ./.dist/app
+if [ -d ".dist" ];then
+    rm .dist -r
+fi
+mkdir .dist
+CGO_ENABLED=1 GOOS=linux GOARCH=amd64 go build -o ./.dist/app
 cp templates  ./.dist -r
 cp static  ./.dist -r
 cp migrations  ./.dist -r
